refactor(frontend): separate theme registry from theme index
Move all theme imports to index.ts (single file coders edit to add themes).
Registry.ts is now import-free — just the registration API. This eliminates
merge conflicts when multiple coders add themes in parallel.
