+++ frontend/src/themes/index.ts
@@ -1,0 +1,13 @@
+import { registerTheme } from './registry';
+import { oracleDefault } from './oracle-default';
+import { midnightTeal } from './midnight-teal';
+import { emeraldNoir } from './emerald-noir';
+import { sakura } from './sakura';
+
+registerTheme(oracleDefault);
+registerTheme(midnightTeal);
+registerTheme(emeraldNoir);
+registerTheme(sakura);
+
+export { getThemes, getTheme, DEFAULT_THEME_ID } from './registry';
+export type { ThemeDefinition, ThemeTokens } from './types';
